requirements.txt
openenv-core[core]>=0.2.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
numpy>=1.24.0
openai>=1.0.0
pyyaml>=6.0
matplotlib>=3.5.0
httpx>=0.24.0
websockets>=13.0